Move auto-translation into events folder
